requirements.txt
numpy>=1.15.0
Pillow>=5.0.0
torch>=1.6.0
pytorch-hed